+++ sdk/azcore/internal/exported/response_error_test.go
@@ -468,3 +468,8 @@ func TestNilRequestInRawResponse(t *testing.T) {
 	}
 	require.EqualValues(t, expected, respErr.Error())
 }
+
+func TestNilResponseBody(t *testing.T) {
+	const expected = "Request information not available\n--------------------------------------------------------------------------------\nRESPONSE 0: \nERROR CODE UNAVAILABLE\n--------------------------------------------------------------------------------\nResponse contained no body\n--------------------------------------------------------------------------------\n"
+	require.EqualValues(t, expected, (&ResponseError{RawResponse: &http.Response{}}).Error())
+}